add language codes to language picker
As I understand it, this should or could allow selecting these languages
by typing their names in bottom right selector. (They can already be
selected by adding to &language in the URL.)

These are language codes of languages indigenous to Canada I saw in
https://taginfo.geofabrik.de/north-america:canada/keys that have over 50
usages. This is intended as a best effort start, not an exclusive list,
and I'd be happy to add others.